Making Nixy more generic.
Tasks is now a new struct similar to the Marathon Tasks API.
This adds new values to the template such as StartedAt, StagedAt, Version, ServicePorts and Ports is now an array to support tasks running on multiple ports.

We have also added PortDefinitions to the Apps, with Port, Protocol and Labels.
This is added for future support of UDP/TCP traffic proxying.