Support for Marathon's `portMapping` config
Marathon's `portDefinition` config is only used for apps running with
host-mode networking. Since Marathon 1.5, it will no longer include
port-definition configuration in its API output if it is not in use,
which means that it is no longer possible for Nixy to read port-level
labels from the Marathon API.

This commit adds access to Marathons `app.container.portMappings` 
property to Nixy's templates, allowing us to access this configuration.